move addPost to containers as newpost
